+++ CHANGES.md
@@ -5,6 +5,7 @@
 + More permissive projection application: for example `+[][2;3]` is valid, even
   though `+[]` has rank 1 when used in adverbial contexts that make use of
   function rank.
++ For consistency, make run s form returns its standard output too.
 
 # v0.10.0 2023-03-26